refactor: change project structure to follow go standard layout
