+++ internal/cron-jobs/cronjob_integration_test.go
@@ -6,11 +6,11 @@ import (
 
 	"github.com/DATA-DOG/go-sqlmock"
 	"github.com/golang/mock/gomock"
-	cronjobs "github.com/skobelina/currency_converter/domains/cron-jobs"
-	"github.com/skobelina/currency_converter/domains/mails"
-	"github.com/skobelina/currency_converter/domains/rates"
-	"github.com/skobelina/currency_converter/domains/subscribers"
-	"github.com/skobelina/currency_converter/repo"
+	cronjobs "github.com/skobelina/currency_converter/internal/cron-jobs"
+	"github.com/skobelina/currency_converter/internal/mails"
+	"github.com/skobelina/currency_converter/internal/rates"
+	"github.com/skobelina/currency_converter/internal/subscribers"
+	"github.com/skobelina/currency_converter/pkg/repo"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"gorm.io/gorm"
