+++ internal/mails/templates/exchange-rate.go
