+++ internal/mails/templates/styles.go
