+++ internal/rates/handler.go
@@ -6,9 +6,9 @@ import (
 	"github.com/gorilla/mux"
 	"gorm.io/gorm"
 
-	utils "github.com/skobelina/currency_converter/utils/errors"
-	"github.com/skobelina/currency_converter/utils/rest"
-	"github.com/skobelina/currency_converter/utils/serializer"
+	utils "github.com/skobelina/currency_converter/pkg/utils/errors"
+	"github.com/skobelina/currency_converter/pkg/utils/rest"
+	"github.com/skobelina/currency_converter/pkg/utils/serializer"
 )
 
 type RateServiceInterface interface {
