+++ internal/rates/model.go
