+++ pkg/utils/rest/registrable-service.go
